UI/requirements.txt
streamlit
pandas
google-auth
google-auth-httplib2
google-cloud-aiplatform
google-generativeai
numpy
joblib
nltk
scikit-learn
matplotlib
streamlit
pandas
protobuf
python-dotenv
